Get the OSM network functions

Get the OSM network functions

Usage

get_traffic_calming(osm_points, filter = TRUE)

Arguments

osm_points

An sf object with the OSM network

filter

Whether to filter the results to only include traffic calming measures

Value

A data frame with the OSM network relevant for active travel

Examples

osm_points = get_points("Edinburgh")
#> No exact match found for place = Edinburgh and provider = geofabrik. Best match is Limburg. 
#> Checking the other providers.
#> An exact string match was found using provider = bbbike.
#> Downloading the OSM extract:
#> Error: The download operation was aborted. If this was not intentional, you may want to increase the timeout for internet operations to a value >= 300 by using options(timeout = ...) before re-running this function. We also suggest you to remove the partially downloaded file by running the following code (possibly in a new R session): file.remove("/tmp/Rtmp4I8qlF/bbbike_Edinburgh.osm.pbf")
traffic_calming = get_traffic_calming(osm_points)
#> Error: object 'osm_points' not found
head(traffic_calming)
#> Error: object 'traffic_calming' not found
plot(traffic_calming["traffic_calming"])
#> Error: object 'traffic_calming' not found