Long format Level of Service (LOS) table

This dataset contains the level of service information in a long format, with columns for speed limit, AADT, infrastructure type, and level of service.

Usage

data(los_table_long)

Format

A data frame with columns including speed limit, AADT, cycle_segregation and level_of_service

Source

Generated from los_table_npt dataset

Examples

data(los_table_long)
head(los_table_long)
#> # A tibble: 6 × 5
#>   `Speed (85th kph)` `Speed Limit (mph)` AADT    infrastructure level_of_service
#>   <chr>              <chr>               <chr>   <chr>                     <dbl>
#> 1 30 kph to 50 kph   20 mph              0 to 1… Mixed Traffic…                3
#> 2 30 kph to 50 kph   20 mph              0 to 1… Off road cycl…                3
#> 3 30 kph to 50 kph   20 mph              0 to 1… Painted Cycle…                3
#> 4 30 kph to 50 kph   20 mph              0 to 1… Segregated Tr…                3
#> 5 30 kph to 50 kph   20 mph              0 to 1… Segregated Tr…                3
#> 6 30 kph to 50 kph   20 mph              2000 t… Mixed Traffic…                2